requirements.txt
torch
torchvision
pandas
scikit-learn
tqdm
Pillow